requirements.txt
# Minimal project requirements. Install torch separately to match your CUDA version:
# e.g. pip install torch --index-url https://download.pytorch.org/whl/cu128
ultralytics>=8.3.228
tensorflow>=2.13.0
opencv-python-headless>=4.7.0
streamlit>=1.20
numpy>=1.24
pillow>=9.0
matplotlib>=3.6
albumentations>=1.3.0
tqdm>=4.65.0
pandas>=1.5.0
plotly>=5.0.0
pyyaml>=6.0
# --- IGNORE ---
